# openapi.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AdminResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "example": "ADMIN"
          },
          "isActive": {
            "type": "boolean"
          },
          "branchId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "branch": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateAdminRequest": {
        "type": "object",
        "required": [
          "fullName",
          "username",
          "password",
          "branchId"
        ],
        "properties": {
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "username": {
            "type": "string",
            "example": "john_admin"
          },
          "password": {
            "type": "string",
            "minLength": 10,
            "example": "secret12345"
          },
          "branchId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "UpdateAdminRequest": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "branchId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
  "openapi": "3.0.0",
  "info": {
    "title": "Store Management API",
    "version": "1.0.0",
    "description": "Multi-branch retail management system API"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/admins": {
      "post": {
        "summary": "Create a new admin and assign to a branch",
        "tags": [
          "Admins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admin created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResponse"
                }
              }
            }
          },
          "404": {
            "description": "Branch not found"
          },
          "409": {
            "description": "Username already taken"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "summary": "List all admins (filterable by branch and status)",
        "tags": [
          "Admins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of admins"
          }
        }
      }
    },
    "/admins/{id}": {
      "get": {
        "summary": "Get a single admin by ID",
        "tags": [
          "Admins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin found"
          },
          "404": {
            "description": "Admin not found"
          }
        }
      },
      "patch": {
        "summary": "Update admin details or reassign to another branch",
        "tags": [
          "Admins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin updated"
          },
          "404": {
            "description": "Admin or branch not found"
          }
        }
      },
      "delete": {
        "summary": "Permanently delete an admin",
        "tags": [
          "Admins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Admin deleted"
          },
          "404": {
            "description": "Admin not found"
          }
        }
      }
    },
    "/admins/{id}/disable": {
      "patch": {
        "summary": "Disable an admin account",
        "tags": [
          "Admins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin disabled"
          }
        }
      }
    },
    "/admins/{id}/enable": {
      "patch": {
        "summary": "Re-enable a disabled admin account",
        "tags": [
          "Admins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin enabled"
          }
        }
      }
    },
...
  },
...
}